Split transaction records and events into purchases and sales
For the sake of long-term maintainability and room for expansion, record and event classes denoting a transaction have been made abstract,
and purchase/sale sub-classes are now instantiated in their parents place, denoting the type of transaction without the use of boolean values.

In addition, sales now can be marked as a listing, in order to directly accomidate Broker implementations which put objects up for sale rather than immediately sell them.

Version has been increased to 1.0.0-rc1, as this update is incompatible with prior versions but is expected to be the final form of the API for the foreseeable future.